Manikanta/00_Core_Foundation_Completed/apps/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
pyjwt>=2.8.0
cryptography>=41.0.0
httpx>=0.25.0
slowapi>=0.1.9
clerk-sdk>=1.0.0

